Battery code working better.
It appears that there is some conflict between the battery code and the
wifi/bluetooth code.  I suspect that reading from pin 29 is causing
though i have no idea (yet) why.

May need to partition the code so that the networking code reinitializes
every time it is executed, and the same for bluetooth and battery.